# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Conferência de Pedidos API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/logs": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "Listar Logs",
        "operationId": "listar_logs_api_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "data_ini",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Ini"
            }
          },
          {
            "name": "data_fim",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Fim"
            }
          },
          {
            "name": "tipo_acao",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro 'contém' no tipo de ação",
              "title": "Tipo Acao"
            },
            "description": "Filtro 'contém' no tipo de ação"
          },
          {
            "name": "usuario",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Usuario"
            }
          },
          {
            "name": "pedido_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pedido Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/limpar-retencao": {
      "post": {
        "tags": [
          "logs"
        ],
        "summary": "Limpar Retencao",
        "operationId": "limpar_retencao_api_logs_limpar_retencao_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LimpezaRetencaoRequest",
                "default": {
                  "dias_logs": 30,
                  "dias_recebimentos": 120
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/etiquetas-prefetch/status": {
      "get": {
        "tags": [
          "etiquetas-prefetch"
        ],
        "summary": "Obter Status",
        "operationId": "obter_status_api_etiquetas_prefetch_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/etiquetas-prefetch/config": {
      "get": {
        "tags": [
          "etiquetas-prefetch"
        ],
        "summary": "Obter Config",
        "operationId": "obter_config_api_etiquetas_prefetch_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "etiquetas-prefetch"
        ],
        "summary": "Salvar Config",
        "operationId": "salvar_config_api_etiquetas_prefetch_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigBody": {
        "properties": {
          "usuarios": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Usuarios"
          }
        },
        "type": "object",
        "title": "ConfigBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LimpezaRetencaoRequest": {
        "properties": {
          "dias_logs": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 7.0,
            "title": "Dias Logs",
            "description": "Apaga logs mais antigos que N dias",
            "default": 30
          },
          "dias_recebimentos": {
            "type": "integer",
            "maximum": 730.0,
            "minimum": 30.0,
            "title": "Dias Recebimentos",
            "description": "Apaga cargas de recebimento (timeline) mais antigas que N dias",
            "default": 120
          }
        },
        "type": "object",
        "title": "LimpezaRetencaoRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}